Fix a bug on Ubuntu with GlobalMenu.
Activate signal was not catched fot first GtkItem. So the sensitivity of MenuItem are set on loading of images and sequences and (for Undo/Redo) when undo operations are executed.

The function update_MenuItem MUST be written in the right place each time a new function loading data on screen is coded.